call sage-connector once, move func back to tasks
