+++ apps/sage_intacct/models.py
@@ -5,16 +5,15 @@ from datetime import datetime
 from django.conf import settings
 from django.db.models import Q,JSONField
 from django.db import models
+from django.utils.module_loading import import_string
 
-
-from apps.sage_intacct.import_helpers import get_or_create_credit_card_vendor
 from fyle_accounting_mappings.models import Mapping, MappingSetting, DestinationAttribute, CategoryMapping, \
     EmployeeMapping
 
 from apps.fyle.models import ExpenseGroup, Expense, ExpenseAttribute, Reimbursement, ExpenseGroupSettings, DependentFieldSetting
 from apps.mappings.models import GeneralMapping
 
-from apps.workspaces.models import Configuration, Workspace, FyleCredential
+from apps.workspaces.models import Configuration, Workspace, FyleCredential, SageIntacctCredential
 from typing import Dict, List, Union
 
 
@@ -874,9 +873,9 @@ class JournalEntryLineitem(models.Model):
 
     class Meta:
         db_table = 'journal_entry_lineitems'
-    
+
     @staticmethod
-    def create_journal_entry_lineitems(expense_group: ExpenseGroup, configuration: Configuration):
+    def create_journal_entry_lineitems(expense_group: ExpenseGroup, configuration: Configuration, sage_intacct_connection):
         """
         Create journal entry lineitems
         :param expense_group: expense group
@@ -937,7 +936,7 @@ class JournalEntryLineitem(models.Model):
 
                 if lineitem.fund_source == 'CCC' and configuration.use_merchant_in_journal_line:
                     # here it would create a Credit Card Vendor if the expene vendor is not present
-                    vendor = get_or_create_credit_card_vendor(expense_group.workspace_id, configuration, lineitem.vendor)
+                    vendor = import_string('apps.sage_intacct.tasks.get_or_create_credit_card_vendor')(expense_group.workspace_id, configuration, lineitem.vendor, sage_intacct_connection)
                     if vendor:
                         vendor_id = vendor.destination_id
