+++ tests/test_sageintacct/test_tasks.py
@@ -15,7 +15,7 @@ from fyle_accounting_mappings.models import DestinationAttribute, EmployeeMappin
 from apps.fyle.models import ExpenseGroup, Reimbursement
 from apps.sage_intacct.utils import SageIntacctConnector
 from apps.mappings.models import GeneralMapping
-from apps.sage_intacct.import_helpers import get_or_create_credit_card_vendor
+from apps.sage_intacct.tasks import get_or_create_credit_card_vendor
 from .fixtures import data
 from django.conf import settings
